Finish first test of regularization code
It runs without error and produces reasonable results. One change
needed to make it robust was to do a 1D optimization to find a good
starting piont for the intercept estimate. Very poor initial values
cause the search to veer out into the extremes of parameter space,
which leads to problems calculating the likelihood. However, a
starting point very close to the optimum seems to cause poor
performance for some reason. A very long loop of "insufficent
decreases" and step size adjustments ensues. This was remedied by
increasing the tolerance of the 1D optimizer so that it produces a
very crude estimate. However, this is not a complete fix and something
more complete should be done eventually.